+++ src/client/hooks/vue/index.ts
@@ -1,4 +1,4 @@
-export * from './useOnStandalonePluginInitVue';
-export * from './usePluginSettingsVue';
-export * from './usePluginStateVue';
-export * from './useWatchPluginMessageVue';
+export * from './useOnStandalonePluginInit';
+export * from './usePluginSettings';
+export * from './usePluginState';
+export * from './useWatchPluginMessage';
