+++ src/client/hooks/vue/usePluginState.ts
@@ -1,12 +1,14 @@
-import { UnwrapRef, onMounted, ref, watch } from 'vue';
+import { UnwrapRef, ref, watch, watchEffect } from 'vue';
 
 import { lobeChat } from '@/client';
 
-export const usePluginSettingsVue = <T>(initialValue: T) => {
+export const usePluginState = <T>(key: string, initialValue: T) => {
   const value = ref(initialValue);
-  onMounted(() => {
-    lobeChat.getPluginSettings().then((e) => {
+
+  watchEffect(() => {
+    lobeChat.getPluginState(key).then((e) => {
       if (!e) return;
+
       value.value = e;
     });
   });
@@ -18,9 +20,9 @@ export const usePluginSettingsVue = <T>(initialValue: T) => {
   watch(
     () => value.value,
     (newValue) => {
-      lobeChat.setPluginSettings(newValue);
+      lobeChat.setPluginState(key, newValue);
     },
   );
 
-  return [value, updateValue] as const;
+  return [value.value, updateValue] as const;
 };
